Fixed regression introduced in r9027.
Review URL: http://codereview.chromium.org/7741041

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@9029 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/x64/regexp-macro-assembler-x64.cc
@@ -1185,7 +1185,7 @@ int RegExpMacroAssemblerX64::CheckStackGuardState(Address* return_address,
   MaybeObject* result = Execution::HandleStackGuardInterrupt();
 
   if (*code_handle != re_code) {  // Return address no longer valid
-    int delta = *code_handle - re_code;
+    intptr_t delta = *code_handle - re_code;
     // Overwrite the return address on the stack.
     *return_address += delta;
   }
